Fix env variable parsing
Env ariables were parsed as strings
due to the use of "_" and "n".
That's only allowed in js/ts code and not
in the .env definition, so I removed those.

I also updated the docker compose file because
I forgot to pass some env variables in the
request_processor service.