Separate structurizr from docker compose
A structurizr container is not part of the
minting server solution, it only provides the docs
so it should not be included in the release that
we provide to the community.

Instead, provide instructions on how to run this separately
so that devs can still interact with the C4 diagrams
if they clone the repo.